Moved OpenSkos Exceptions to separate namespace
